refactor(automations): extract reminder detail panel
