+++ src/components/automations-view.test.ts
@@ -7,16 +7,15 @@ const source = [
   readFileSync(new URL("./automations/status-icon.tsx", import.meta.url), "utf8"),
   readFileSync(new URL("./automations/cron-detail-primitives.tsx", import.meta.url), "utf8"),
   readFileSync(new URL("./automations/cron-detail-panel.tsx", import.meta.url), "utf8"),
+  readFileSync(new URL("./automations/reminder-detail-panel.tsx", import.meta.url), "utf8"),
   readFileSync(new URL("./automations/schedule-list.tsx", import.meta.url), "utf8"),
   readFileSync(new URL("./automations/inbox-feed-list.tsx", import.meta.url), "utf8"),
   readFileSync(new URL("./automations/templates-panel.tsx", import.meta.url), "utf8"),
   readFileSync(new URL("./automations/ritual-overview.tsx", import.meta.url), "utf8"),
 ].join("\n");
 const codexDetailPanel = readFileSync(new URL("./automations/cron-detail-panel.tsx", import.meta.url), "utf8");
-const detailPanelControls = source.slice(
-  source.indexOf("function DetailPanel"),
-  source.indexOf("function NewMenuItem"),
-);
+const reminderDetailPanel = readFileSync(new URL("./automations/reminder-detail-panel.tsx", import.meta.url), "utf8");
+const detailPanelControls = reminderDetailPanel.slice(reminderDetailPanel.indexOf("export function DetailPanel"));
 
 assert.match(source, /from "@\/components\/automations\/templates-panel"/, "template browsing is owned by its dedicated module");
 assert.match(source, /from "@\/components\/automations\/ritual-overview"/, "Ritual overview rows and calendar helpers are owned by a dedicated module");
